requirements.txt
Flask
requests
python-telegram-bot
threading
numpy
pandas
scikit-learn
joblib
gunicorn